move some firebase stuff into App.js
