+++ src/App.js
@@ -1,6 +1,29 @@
 import logo from './logo.svg';
 import './App.css';
 import Routing from "./routes/index";
+import { getFirestore, collection, getDocs } from 'firebase/firestore';
+import { initializeApp } from "firebase/app";
+import { getAnalytics } from "firebase/analytics";
+
+
+
+const firebaseConfig = {
+  apiKey: "AIzaSyCo0gPK4fEqqPd5WgmhISKfy1P8AgPmf3Y",
+  authDomain: "purduezilla.firebaseapp.com",
+  databaseURL: "https://purduezilla-default-rtdb.firebaseio.com",
+  projectId: "purduezilla",
+  storageBucket: "purduezilla.appspot.com",
+  messagingSenderId: "1043168052788",
+  appId: "1:1043168052788:web:be5443766ed89f9479ac61",
+  measurementId: "G-9G4NCBXCZX"
+};
+
+const fireapp = initializeApp(firebaseConfig);
+const analytics = getAnalytics(fireapp);
+
+const db = getFirestore(fireapp);
+
+
 
 function App() {
   return (
@@ -11,4 +34,4 @@ function App() {
   );
 }
 
-export default App;
+export default { App, fireapp, analytics, db };
